Refactor: Loan Registration File Structure
