+++ src/components/tabs/collateral-management-tab.tsx
