+++ src/components/tabs/loan-calculation-tab.tsx
